+++ crates/api-model/src/rpc_conv/mod.rs
@@ -27,7 +27,16 @@ pub mod dhcp_record;
 pub mod dns;
 pub mod dpa_interface;
 pub mod dpu_remediation;
+pub mod expected_machine;
+pub mod expected_power_shelf;
+pub mod expected_rack;
+pub mod expected_switch;
+pub mod extension_service;
+pub mod hardware_info;
+pub mod health;
+pub mod ib_partition;
 pub mod instance;
+pub mod instance_type;
 pub mod operating_system_definition;
 pub mod os;
 pub mod power_manager;
